# Conflicts:
#   ForwardDetectors/ZDC/ZdcAnalysis/Root/ZDCDataAnalyzer.cxx
#   ForwardDetectors/ZDC/ZdcAnalysis/Root/ZDCFitWrapper.cxx
#   ForwardDetectors/ZDC/ZdcAnalysis/Root/ZDCPulseAnalyzer.cxx
#   ForwardDetectors/ZDC/ZdcAnalysis/Root/ZdcAnalysisTool.cxx
#   ForwardDetectors/ZDC/ZdcAnalysis/Root/ZdcSincInterp.cxx
#   ForwardDetectors/ZDC/ZdcAnalysis/ZdcAnalysis/ZDCDataAnalyzer.h
#   ForwardDetectors/ZDC/ZdcAnalysis/ZdcAnalysis/ZDCFitWrapper.h
#   ForwardDetectors/ZDC/ZdcAnalysis/ZdcAnalysis/ZDCPulseAnalyzer.h
#   ForwardDetectors/ZDC/ZdcAnalysis/ZdcAnalysis/ZdcAnalysisTool.h
#   ForwardDetectors/ZDC/ZdcAnalysis/ZdcAnalysis/ZdcSincInterp.h

+++ TestBeam/TBRec/src/TBTrackInfoFromTag.cxx
@@ -108,7 +108,7 @@ TBTrackInfoFromTag::execute()
 	ATH_CHECK( evtStore()->retrieve(evtInfo,m_SGkey1) );
 
 	int runNumber = evtInfo->event_ID()->run_number();
-	int evtNumber = evtInfo->event_ID()->event_number();
+	uint64_t evtNumber = evtInfo->event_ID()->event_number();
 
 	ATH_MSG_DEBUG("Run/Event numbers:\t " << runNumber << " / " << evtNumber);
